requirements.txt
confluent-kafka==1.9.2
Flask==2.2.2
delta-spark==2.1.1
pyspark==3.3.1
requests==2.28.1
avro==1.11.1
inquirer
Click>=6.0
rich
inquirer
dacite
pyyaml
dataclasses_avroschema
streamlit==1.15.2
python-terraform==0.10.1